feat!: make a stroke based undo, breaking win32
I would really like to improve the tracking of strokes. atm we use
undoCount and strokeCount to tell whether I have 0 or 1, but that should
be encoded better. Maybe after many shower thoughts, I can figure out
the gist better. If the reader has a suggestion for that, I would be
happy to consider.